Refactor: renamed 'fixtures' module to 'pga_fixtures' for better clarity and consistency in tests. - Adjusted module imports accordingly.
